refactor: move unit tests from lib.rs to tests/unit.rs
- Move 38 unit tests to tests/unit.rs for better organization
- Make format_inline, strip_ansi, extract_href, wrap_text public
- Tests now run as integration tests alongside conformance tests
